+++ src/main/kotlin/com/filip/machaj/demo/repo/ModelRepo.kt
@@ -7,7 +7,7 @@ import com.filip.machaj.demo.model.dane.Model
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.CrudRepository
 
-interface ModelRepo: CrudRepository<Model,Long> {
+interface ModelRepo: CrudRepository<Model,Int> {
 
     @Query("select * from model where id=( select max(id) from model)", nativeQuery = true )
     @JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator::class, property = "id")
